requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
sentence-transformers==2.2.2
chromadb==0.4.14
PyPDF2==3.0.1
python-docx==0.8.11
python-multipart==0.0.6
torch==2.2.0
tqdm==4.66.1
numpy==1.26.0
pydantic==2.3.0